+++ jeap-messaging-avro-maven-plugin/src/test/java/ch/admin/bit/jeap/messaging/avro/plugin/git/GitClientIT.java
@@ -1,6 +1,7 @@
 package ch.admin.bit.jeap.messaging.avro.plugin.git;
 
 import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.logging.SystemStreamLog;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
@@ -11,21 +12,21 @@ class GitClientIT {
 
     @Test
     void diffCommit() throws MojoExecutionException {
-        GitClient gitClient = new GitClient("/home/dev/development/projects/jme-message-type-registry", "master");
+        GitClient gitClient = new GitClient("/home/dev/development/projects/jme-message-type-registry", "http://mock", "master", new SystemStreamLog());
         final GitDiffDto gitDiff = gitClient.getGitDiff("feature");
         assertNotNull(gitDiff);
     }
 
     @Test
     void getDiffFromLastTag() throws MojoExecutionException {
-        GitClient gitClient = new GitClient("/home/dev/development/projects/jme-message-type-registry", "master");
+        GitClient gitClient = new GitClient("/home/dev/development/projects/jme-message-type-registry", "http://mock", "master", new SystemStreamLog());
         final GitDiffDto gitDiff = gitClient.getDiffFromLastTag();
         assertNotNull(gitDiff);
     }
 
     @Test
     void getDiffFromMaster() throws MojoExecutionException {
-        GitClient gitClient = new GitClient("/home/dev/development/projects/jme-message-type-registry", "master");
+        GitClient gitClient = new GitClient("/home/dev/development/projects/jme-message-type-registry", "http://mock", "master", new SystemStreamLog());
         final GitDiffDto gitDiff = gitClient.getDiffFromTrunk();
         assertNotNull(gitDiff);
     }