+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.50.0] - 2025-07-02
+
+### Added
+
+- Support for adding the tags from the remote repository in the avro-maven-plugin
+- Support for Git token authentication when accessing a remote repository in the avro-maven-plugin
+- Support for using the system's Git installation when accessing a remote repository without a Git token in the avro-maven-plugin 
+
 ## [8.49.1] - 2025-06-30
 
 ### Fixed